+++ mt19937ar.h
@@ -8,7 +8,8 @@ extern "C" {
 #define MT19937AR_VERSION "0.20100917"
 
 void mt_init_genrand(unsigned long s);
-double mt_genrand_res53(void);
+
+float mt_genrand_res53(void);
 
 #ifdef __cplusplus
 }
